requirements.txt
flask
cmake
face_recognition
gunicorn
tensorflow-cpu